MyResAI/requirements.txt
langchain
python-dotenv
openai
sarvamai
chromadb
pytest
SpeechRecognition
PyAudio
fastapi
uvicorn[standard]
requests
httpx
